pyproject.toml
[project]
name = "crypto_trading_bot"
version = "0.1.0"
description = "Modular crypto trading bot with FastAPI backend and machine learning."
readme = "README.md"
requires-python = ">=3.12.10, <3.13"
authors = [
  { name = "Ovie", email = "oviemsaniyo30@gmail.com" }
]

dependencies = [
  "aiohappyeyeballs==2.6.1",
  "aiohttp==3.12.13",
  "aiosignal==1.3.2",
  "alembic==1.16.4",
  "altair==5.5.0",
  "annotated-types==0.7.0",
  "anyio==4.9.0",
  "attrs==25.3.0",
  "bcrypt==4.3.0",
  "blinker==1.9.0",
  "cachetools==5.5.2",
  "certifi==2025.6.15",
  "cffi==1.17.1",
  "charset-normalizer==3.4.2",
  "click==8.2.1",
  "cloudpickle==3.1.1",
  "colorama==0.4.6",
  "contourpy==1.3.2",
  "cryptography==45.0.5",
  "cycler==0.12.1",
  "dateparser>=1.1.8,<2.0",   # relaxed pin
  "pytz>=2024.1,<2025.0",     # relaxed pin
  "dnspython==2.7.0",
  "docker==7.1.0",
  "ecdsa==0.19.1",
  "email-validator==2.2.0",
  "entrypoints==0.4",
  "fastapi==0.111.0",
  "fastapi-cli==0.0.8",
  "fonttools==4.59.0",
  "frozenlist==1.7.0",
  "gitdb==4.0.12",
  "gitpython==3.1.44",
  "graphene==3.4.3",
  "graphql-core==3.2.6",
  "graphql-relay==3.2.0",
  "greenlet==3.2.3",
  "h11==0.16.0",
  "httpcore==1.0.9",
  "httptools==0.6.4",
  "httpx==0.28.1",
  "idna==3.10",
  "imbalanced-learn==0.12.4",
  "importlib-metadata==7.2.1",
  "iniconfig==2.1.0",
  "itsdangerous==2.2.0",
  "jinja2==3.1.6",
  "joblib==1.5.1",
  "jsonschema==4.25.0",
  "jsonschema-specifications==2025.4.1",
  "kaleido==0.2.1",
  "kiwisolver==1.4.8",
  "mako==1.3.10",
  "markdown==3.8.2",
  "markdown-it-py==3.0.0",
  "markupsafe==3.0.2",
  "mdurl==0.1.2",
  "mlflow==2.13.0",
  "multidict==6.5.1",
  "narwhals==1.42.0",
  "numpy==1.26.4",
  "opentelemetry-api==1.35.0",
  "opentelemetry-sdk==1.35.0",
  "opentelemetry-semantic-conventions==0.56b0",
  "orjson==3.11.0",
  "packaging==24.0",
  "pandas==2.3.0",
  "passlib==1.7.4",
  "pillow==10.4.0",
  "pip==25.1.1",
  "pip-tools==7.4.1",
  "plotly==6.1.2",
  "pluggy==1.6.0",
  "propcache==0.3.2",
  "protobuf==4.25.8",
  "psycopg2-binary==2.9.9",
  "pyarrow==15.0.2",
  "pyasn1==0.6.1",
  "pycparser==2.22",
  "pycryptodome==3.23.0",
  "pydantic==2.11.7",
  "pydantic-core==2.33.2",
  "pydeck==0.9.1",
  "pygments==2.19.2",
  "pyparsing==3.2.3",
  "pytest==8.4.1",
  "pytest-mock==3.14.1",
  "pytest-asyncio==1.2.0",
  "python-binance==1.0.29",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.1.1",
  "python-jose==3.3.0",
  "python-multipart==0.0.20",
  "python-telegram-bot==22.5",
  "pywin32==311",
  "pyyaml==6.0.2",
  "querystring-parser==1.2.4",
  "referencing==0.36.2",
  "regex==2024.11.6",
  "requests==2.32.4",
  "rich==13.9.4",
  "rich-toolkit==0.14.8",
  "rpds-py==0.26.0",
  "rsa==4.9.1",
  "scikit-learn==1.3.2",
  "scipy==1.15.3",
  "shellingham==1.5.4",
  "six==1.17.0",
  "smmap==5.0.2",
  "sniffio==1.3.1",
  "sqlalchemy==2.0.41",
  "sqlparse==0.5.3",
  "starlette==0.37.2",
  "streamlit==1.35.0",
  "ta==0.11.0",
  "tenacity==8.5.0",
  "threadpoolctl==3.6.0",
  "toml==0.10.2",
  "tornado==6.5.1",
  "typer==0.16.0",
  "typing-extensions==4.14.1",
  "typing-inspection==0.4.1",
  "tzdata==2025.2",
  "tzlocal==5.3.1",
  "ujson==5.10.0",
  "urllib3==2.5.0",
  "uvicorn==0.35.0",
  "waitress==3.0.2",
  "watchdog==6.0.0",
  "watchfiles==1.1.0",
  "websockets==15.0.1",
  "werkzeug==3.1.3",
  "xgboost==3.0.2",
  "yarl==1.20.1",
  "zipp==3.23.0",
  "backoff==2.2.1",
  "arch>=6.2.0",
  "statsmodels>=0.14.0",
  "hmmlearn>=0.3.0",
  "pykalman>=0.9.5",
  "tsfresh>=0.20.0"
]

[build-system]
requires = ["uv", "setuptools"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "pytest-mock"]
